Create es-modular-ship-expander
Update to v0.0.3
This is a plugin for the game Endless Sky that adds several Modular Outfits for ships. The outfits allow increases to outfit, cargo, bunk, weapon, engine, hull, or shields respectively to an existing ship. There is also an Android Outfit included. All outfits are balanced in various ways.